feat(formula) : 完成配方相关接口
Description:
   完成配方相关接口并完整增改单元测试

+++ backend/src/main/java/com/sophia/store/entity/po/BasicFormula.java
@@ -36,8 +36,8 @@ public class BasicFormula implements Serializable {
     private Integer id;
     @Column(name = "count", nullable = false)
     @ApiModelProperty(value = "数量")
-    private Float count;
-    @OneToOne(fetch = FetchType.EAGER, cascade={CascadeType.MERGE, CascadeType.REFRESH})
+    private Integer count;
+    @OneToOne(fetch = FetchType.EAGER, cascade = {CascadeType.MERGE, CascadeType.REFRESH})
     @JoinColumn(name = "basic_id", referencedColumnName = "id")
     @ApiModelProperty(value = "基础产品")
     private Basic basic;
